requirements.txt
Flask==3.0.0

flasgger~=0.9.7.1
marshmallow~=3.20.1
vyper-config~=1.2.1